requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
langchain_google_genai
wikipedia
langchainhub
groq
langchain-groq
beautifulsoup4
selenium
unstructured